pyproject.toml
[project]
name = "social-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.129.0",
    "fastapi-users[sqlalchemy]>=15.0.4",
    "imagekitio>=5.2.0",
    "python-multipart>=0.0.9",
    "requests>=2.31.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.55.0",
    "uvicorn[standard]>=0.41.0",
]
